pkg/statistics/statistics.go
package statistics

import (
	"time"
)

type Statistic struct {
	Type        string
	Identifier  string
	Count       int
	LastUpdated time.Time
}

type StatisticsService struct {
	// This would be your connection to Redis or any other database you're using
	db Database
}

func NewStatisticsService(db Database) *StatisticsService {
	return &StatisticsService{db: db}
}

func (s *StatisticsService) GetStatistics(statType string, period string, limit int) ([]Statistic, error) {
	// Here you would implement the logic to get statistics from the database
	// based on the type, period, and limit
}

func (s *StatisticsService) SubmitStatistic(stat Statistic) error {
	// Here you would implement the logic to submit a new statistic to the database
}